+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.12.0-rc.29",
+  sdkVersion: "4.12.0-rc.30",
   genVersion: "2.918.1",
-  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.29 2.918.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.12.0-rc.30 2.918.1 2.0 @orq-ai/node",
 } as const;