+++ packages/orq-rc/RELEASES.md
@@ -10839,3 +10839,13 @@ Based on:
 - [typescript v4.12.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.29 - packages/orq-rc
+
+## 2026-07-15 12:25:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.1 (2.918.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.30 - packages/orq-rc